Fix small things and address comments
